+++ packages/bridge-remote/src/RemoteBridgeClientAuthenticator.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evolve\Bridge\Remote;
+
+use Psr\Http\Message\RequestInterface;
+
+/**
+ * @experimental
+ */
+interface RemoteBridgeClientAuthenticator
+{
+    /**
+     * @return array<string, list<string>>
+     */
+    public function authenticationHeaders(
+        RequestInterface $request,
+        RemoteBridgeInvocation $invocation,
+    ): array;
+}